Bug 1767063 - make PerformanceEntry.mEpoch narrowing conversion explicit. r=sefeng
The converted value gets passed into TimingNotification which assigns the value
into a PerformanceEntry (this is the only usage). Since PerformanceEntry is
defined in WebIDL, we could not change its types (which is double for mEpoch) so
we were forced to safely convert the value where we did.

I don't think the existing conversion code had any bugs since we converted a
64-bit signed integer timestamp into uint64_t (safe) into a double (which is
safe for the reasons mentioned in the code comments).

Differential Revision: https://phabricator.services.mozilla.com/D145142

UltraBlame original commit: c43eaaf2d26963428337e6147699e0cf9046fd73

+++ dom/performance/PerformanceMainThread.cpp
@@ -405,7 +405,7 @@ void PerformanceMainThread::InsertUserEntry(PerformanceEntry* aEntry) {
   MOZ_ASSERT(NS_IsMainThread());
 
   nsAutoCString uri;
-  uint64_t markCreationEpoch = 0;
+  double markCreationEpoch = 0;
 
   if (StaticPrefs::dom_performance_enable_user_timing_logging() ||
       StaticPrefs::dom_performance_enable_notify_performance_timing()) {
@@ -419,7 +419,11 @@ void PerformanceMainThread::InsertUserEntry(PerformanceEntry* aEntry) {
       
       uri.AssignLiteral("none");
     }
-    markCreationEpoch = static_cast<uint64_t>(PR_Now() / PR_USEC_PER_MSEC);
+
+    
+    
+    
+    markCreationEpoch = static_cast<double>(PR_Now() / PR_USEC_PER_MSEC);
 
     if (StaticPrefs::dom_performance_enable_user_timing_logging()) {
       Performance::LogEntry(aEntry, uri);
